# api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * A³ shopping platform storefront and order API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: string;
  label: string;
  count: number;
}

export interface Product {
  id: string;
  title: string;
  description: string;
  category: string;
  price: number;
  compareAtPrice: number;
  image: string;
  rating: number;
  reviewCount: number;
  badge: string;
  inStock: boolean;
  quantityAvailable?: number;
}

export type StorefrontSummaryPromo = {
  eyebrow: string;
  title: string;
  description: string;
  code: string;
};

export interface StorefrontSummary {
  categories: Category[];
  featured: Product[];
  promo: StorefrontSummaryPromo;
}

export interface CustomerInput {
  email: string;
  /** @minLength 2 */
  name: string;
  phone?: string;
}

export interface CustomerProfileInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 7 */
  phone: string;
  /** @minLength 3 */
  addressLine1: string;
  addressLine2?: string;
  /** @minLength 2 */
  city: string;
  /** @minLength 2 */
  state: string;
  /** @minLength 3 */
  postalCode: string;
}

export type CustomerAddress = {
  addressLine1: string;
  addressLine2?: string;
  city: string;
  state: string;
  postalCode: string;
};

export interface Customer {
  id: string;
  customerNumber: string;
  email: string;
  name: string;
  phone: string;
  address: CustomerAddress;
  createdAt: string;
}

export interface CartLineInput {
  productId: string;
  /**
     * @minimum 1
     * @maximum 20
     */
  quantity: number;
}

export interface CartLineUpdate {
  /**
     * @minimum 0
     * @maximum 20
     */
  quantity: number;
}

export interface CartLine {
  id: string;
  product: Product;
  quantity: number;
  lineTotal: number;
}

export interface Cart {
  id: string;
  customerId: string;
  lines: CartLine[];
  subtotal: number;
  shipping: number;
  total: number;
  itemCount: number;
}

export interface CheckoutInput {
  /** @minLength 3 */
  addressLine1: string;
  addressLine2?: string;
  /** @minLength 2 */
  city: string;
  /** @minLength 2 */
  state: string;
  /** @minLength 3 */
  postalCode: string;
  note?: string;
  paymentMethod?: string;
}

export interface Checkout {
  orderId: string;
  checkoutUrl: string;
  status: string;
}

export interface Order {
  id: string;
  orderNumber: string;
  customerId: string;
  status: string;
  paymentStatus: string;
  deliveryStatus: string;
  items: CartLine[];
  total: number;
  /** @nullable */
  trackingNumber?: string | null;
  createdAt: string;
}

export type ListProductsParams = {
search?: string;
category?: string;
sort?: ListProductsSort;
/**
 * @minimum 1
 * @maximum 48
 */
limit?: number;
};

export type ListProductsSort = typeof ListProductsSort[keyof typeof ListProductsSort];


export const ListProductsSort = {
  featured: 'featured',
  'price-low': 'price-low',
  'price-high': 'price-high',
  rating: 'rating',
} as const;